# Conflicts:
#	appleMarket/target/classes/buyerboard/controller/BuyerboardController.class
#	appleMarket/target/classes/buyerboard/service/BuyerboardServiceImpl.class

+++ appleMarket/src/main/java/member/bean/MemberDTO.java
@@ -35,7 +35,7 @@ public class MemberDTO {
 	private String member_address; 
 	private String member_detailAddr;
 	private int member_gender; 
-	private int member_reputation; 
+	private double member_reputation; 
 	private int member_warning; 
 	private String member_profilePic;
 	private Date member_entryLogtime;
